This commit was generated by cvs2svn to compensate for changes in r16663, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16664 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/georam.c
@@ -378,7 +378,7 @@ int georam_read_snapshot_module(snapshot_t *s)
         goto fail;
 
     if (size > 4096) {
-        log_error(georam_log, "Size %ld in snapshot not supported.", (long)size);
+        log_error(georam_log, "Size %d in snapshot not supported.", (int)size);
         goto fail;
     }
